+++ sshd-common/src/main/java/org/apache/sshd/common/signature/SignatureRSA.java
@@ -78,7 +78,8 @@ public abstract class SignatureRSA extends AbstractSignature {
              *      corresponds to a good-faith implementation and is considered safe to accept.
              */
             String canonicalName = KeyUtils.getCanonicalKeyType(keyType);
-            if (!KeyPairProvider.SSH_RSA.equals(canonicalName) && !KeyPairProvider.SSH_RSA_CERT.equals(canonicalName)) {
+            if ((!KeyPairProvider.SSH_RSA.equals(canonicalName))
+                    && (!KeyPairProvider.SSH_RSA_CERT.equals(canonicalName))) {
                 throw new IllegalArgumentException("Mismatched key type: " + keyType);
             }
             data = encoding.getValue();