+++ sshd-core/src/main/java/org/apache/sshd/server/SshServer.java
@@ -267,6 +267,7 @@ public class SshServer extends AbstractFactoryManager implements ServerFactoryMa
         return hostKeyCertificateProvider;
     }
 
+    @Override
     public void setHostKeyCertificateProvider(HostKeyCertificateProvider hostKeyCertificateProvider) {
         this.hostKeyCertificateProvider = hostKeyCertificateProvider;
     }